Expand datatype for Unlockables, emblems (Medals + nonmedals), and Conditions.
Made sure there is more than enough headroom for our current purposes.
It should be easy to double again if necessary now that the datatypes have been increased... but that would be obscene at this point
- 1024 Unlockables and 1024 Conditions (these were always tied together in slots)
- 2048 emblems (Medals + nonmedals). If we ship with ~250 maps this permits 8 Medals per map - which is higher than we intend right now but could easily fill out in patches

+++ src/k_follower.c
@@ -48,7 +48,7 @@ boolean K_FollowerUsable(INT32 skinnum)
 {
 	// Unlike R_SkinUsable, not netsynced.
 	// Solely used to prevent an invalid value being sent over the wire.
-	UINT8 i;
+	UINT16 i;
 	INT32 fid;
 
 	if (skinnum == -1 || demo.playback)
